set test cases for "collectionsK//vectorK" at "DEBUG.c" (only "test_initfree" and "test_pushpop") and fixing some bugs on "vector_Kutils.c"'s "vectorK_defresize", "vectorK_pushback" and "vectorK_popback"

+++ collectionsK/vectorK/vector_Kutils.h
@@ -10,7 +10,7 @@
 // Vector struct
 typedef struct vectorK {
     dinarrayK* dinarray;    // pointer to the dinarrayK
-    int size;               // size of the vector (number of elements it contains in its data) (must be <= than capacity)
+    size_t size;            // size of the vector (number of elements it contains in its data) (must be <= than capacity)
     int (*resize_func)(int, int);      // pointer to the function for resizing the dinarrayK
 } vectorK;
